Expanded tests for uniq, reduce, every, some
_.uniq had a test for handling iterators that work with a sorted array
that allowed for false positives.  It would pass even if _.uniq wasn't
written to handle iterators.

_.reduce should not skip the first index in the collection when that
value is the same as the starting value of the accumulator.  Added test
that checks for such a scenario.

_.every allowed for a false positive when only the last index is true.
Added a test to account for this case.

Found an implementation of _.some that passed all tests yet returned
false when given an array containing the single value of true.  Added
a test for this.